requirements.txt
beautifulsoup4==4.11.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==2.1.1
click==8.1.7
dnspython==2.7.0
Flask==2.3.3  # 안정적인 호환 버전
Flask-JWT-Extended==4.4.4
flasgger==0.9.5
idna==3.10
itsdangerous==2.1.2  # Flask 2.x와 호환
Jinja2==3.1.2  # Flask 2.x와 호환
MarkupSafe==2.1.3  # Flask 2.x와 호환
mongoengine==0.26.0
numpy==2.1.3
pandas==2.2.3
pymongo==4.3.3
pytest==7.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.28.1
schedule==1.1.0
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==1.26.20
Werkzeug==2.3.7  # Flask 2.x와 호환
